refactor: Rebase + Refactor DAG exporter to working state (#872)
* merge master to dag-exporter (#846)

https://github.com/pinterest/querybook/pull/846

* ui: merge master to dag-exporter (#849)

https://github.com/pinterest/querybook/pull/849

* refactor exporter to working state

* fix lint

* addressed comments
